relabel plot numbers
