+++ figures/plot_S6_S7.sh
